requirements.txt
python-dotenv
browserstack-local
selenium
browserstack-sdk